requirements.dev.txt
django-ninja==1.4.5
joserfc==1.4.1
orjson==3.11.4
